requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-webpush==0.3.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
http_ece==1.2.1
idna==3.10
msgpack==1.1.1
multidict==6.6.4
propcache==0.3.2
py-vapid==1.9.2
pycparser==2.22
PyJWT==2.10.1
pywebpush==2.0.3
redis==6.4.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
